api call fix, core adta saving fix

+++ WeatherApp/Models/WeatherList/CoreDataModels/CityEntity+CoreDataProperties.swift
@@ -17,7 +17,7 @@ extension CityEntity {
     }
 
     @NSManaged public var name: String
-    @NSManaged public var id: Double
+    @NSManaged public var id: Int64
     @NSManaged public var currentWeather: CurrentWeatherEntity?
 
 }
